requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
groq>=0.4.0
google-generativeai>=0.5.0
